Create a table for the library
Created a table to display our library which at the moment is filled
with hard-coded books. JS can populate a table based on what is in
the array of myLibrary.
Next step to bring up the form when the NEW BOOK button is pressed.

From there I will figure out how to add new books into the myLibrary
array as currently I feel, that I would need to know what selectors
to use to do this.