Fix issues around adding delete button
To be able to delete the books we intend, I had to add data attributes.
I also needed a counter on the book objects to have a corresponding
match. Fixed the issue up until the point where deleting the last book
could cause a problem (this is to be tackled next).

Besides adding a counter I also had to solve the issue of printing the
counter while looping through the object keys and values at table
creation.

Then I changed the order of key-value pairs: put the counter as the
first property of the object. This way that can easily be removed from
the beginning of the list as adding it to the end could cause issues in
the future if further properties were added after that.

Furthermore, I moved the addToTable() inside the createTable() to cut
out repetition for ease of maintenance.