Fix remove book issues
Complete revamp on the project.

- now a new Book object is being created every time a book's details
are added.
- removed counter from object (array index does the same thing)
- books are displayed at each round from scratch (as per task) so when
a table has already been created it gets deleted first => this fixes
the issue of button id's not matching
- not putting a data set book id on the row, but on the delete button
instead (hence no longer needed to remove the row as we are recreating
the table instead).
- couldn't add event listener to the delete button, but was able to
add an onclick attribute to it to get the relevant book id.